backend/requirements.txt
django
graphene-django
django-graphql-jwt
django-cors-headers
psycopg2
